pyproject.toml
[project]
name = "ai-agents"
version = "1.0.0"
description = "Google ADK based AI agents with local Ollama and memory support"
authors = [
    {name = "Michael Yu", email = "michaelyu@example.com"},
]
dependencies = [
    "google-adk[a2a]>=1.0.0",
    "litellm>=1.65.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio-throttle>=1.0.0",
    "httpx>=0.28.1",
    "uvicorn>=0.35.0",
    "google-api-python-client>=2.175.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = {file = "LICENSE"}

[project.optional-dependencies]
a2a = ["google-adk[a2a]>=1.0.0"]
mcp = ["mcp>=1.0.0", "google-adk[mcp]>=1.0.0"]
testing = ["requests>=2.31.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "requests>=2.31.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["base-ai-agent", "google-search-agent", "shared", "mcp-tools"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "requests>=2.31.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ai_agents"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
